move contract addresses
